refactor: app.pipeline.parse -> app.parse
